Make it possible give grades higher than ten
This makes it possible for TA's to give students a bonus for certain
assignments. By default this is still not possible, and a maximum grade always
needs to be set. By accident this also fixes a bug that the displayed grade of a
rubric could be higher than the actual grade if this grade was higher than a
ten.