Make it possible to delete submission that have plagiarism cases (#536)
* Make it possible to delete submission that have plagiarism cases

This adds a `ondelete='CASCADE'` to the `file_id` references in the
`PlagiarismMatch` table. This deletes the plagiarism match record when one of
the files is deleted, which is done when deleting submissions.

* Make sure plagiarism cases are not deleted when deleting file

This would enable students to hide plagiarism by simply deleting those
files. This will probably never occur (as the student would need to be able to
alter files after a plagiarism run), however it is better to be safe than sorry.