Remove the option to automatically generate keys (#554)
* Remove the option to automatically generate keys

This automatically generating only worked when the `create_app` function was
called once, as otherwise these keys would be different in different
processes. This approach was very fragile, and would never work for different
servers. This PR removes this functionality and startup fails in production when
the keys are not provided.