Filter users in the user selector (#553)
* Only display users in the course in submission uploader

It was possible to pick users to upload for that weren't in the course, which
resulted in an error when submitting. This fixes #498.

* Make it possible to exclude users from search

This commit also moves the search users in course route to
`/api/v1/courses/id/users/` and it introduces a new permission
`can_list_course_users` which does exactly what you would think it does.