Replace archive dependency with own module (#569)
* Replace archive dependency with own module

This enables us to move some code that manipulate archives into our own
module. This also adds some tests to the actual archive module. This also fixes
a bug that extracted directories could not be removed if the permissions of the
tar file were to restricted.

* Move all processing of directories into archive module

* Fix broken ignore file test

The extension was added to the name when testing without any ignore file,
furthermore the result was not checked when testing without any ignore file.