Refactor lti passback code (#588)
* Refactor lti passback coccde

Refactor the entire code used for lti passback. This is now mostly our code, and
doesn't depend on ``lxml`` anymore. The code is now also actually tested. This
also make it easier to add other LMSsen. Finally this disables the initial
grade passback for canvas starts using the submission details option so that the
hand in date is exactly the same time as in CodeGrade.

* Make sure message identifier is consistent + unique for a submission

* Make sure passback grades is retried if it fails