Replace submitted symlinks with actual files (#627)
* Show a warning when submission contains symlinks.

CodeGrade doesn't support symlinks, but that may break submissions if
for example their build system depends on some symlinks or whatever.
This patch replaces any symlink in submitted archives with a regular
file explaining that there is no symlink support, and with the name the
symlink originally linked to. When a submitted archive contained
symlinks, a warning header is added to the response with a message
detailing which files in the archive were thought to be symlinks. This
warning message is also shown in a popover on the submit button of the
file uploader.

This also removes the warning keyword argument from make_empty_response,
jsonify, and jsonify_extended. You should now use
psef.helpers.add_warning to add warning headers to the response, which
can be called from anywhere.

* Always call maybe_add_warning after each request.