Make it possible to display IPython notebooks (#742)
* Make it possible to display IPython notebooks

This makes it possible to display IPython notebooks and markdown files within
the codeviewer. It is not possible for these file types to be displayed within
the overview mode, that is something left for later.

One thing to consider is that this makes it possible for students to know when
certain files are opened (as external files are loaded). I don't think this is a
very large problem. But maybe somebody else disagrees.