Add a release notifier on the home grid (#787)
* Add a release notifier on the home grid

This makes it clear for users that a new release has been released. The release
commit should have a single alinea explaining the release after the `**
Released**` text. The release note is displayed for 7 days after the `npm
build`. This start date can be overridden using the `CG_FORCE_BUILD_DATE`
environment variable.