Prevent error in console when not logged in on page load (#809)
* Prevent error in console when not logged in on page load

When `verifyLogin` fails it rejects the promise it returns, but the
`created` method in src/main.js did not catch this, so the uncaught
promise rejection was logged to the console.

* Redirect to /login when first verifyLogin fails and page requires login

* Move permissions into the store

This removes all caching functionality for the permissions. The speed improves
for many pages, as it removes the use for one api call.