Make sure order of submissions is stable (#816)
This is needed as `.sort` in javascript is not stable. It is done by comparing
submissions by author when the grade is equal. There is still an edge case when
two submissions are by the same user with the same grade, but this will not
happen often at all. This also makes sure that descending sort is copied from
the submission list page.