Edit snippets in modal (#855)
Snippet replacement texts can be multiple lines, but the user interface didn't reflect this with its single line input. Changing it to a resizeable text area makes it much easier to use for longer replacements. Since there is not a lot of room in the sidebar, clicking the "edit snippet" button now opens a modal.

When the new key of an edited snippet already exists, the user is prompted if they want to override the existing snippet.

Also fixes vertical alignment of icons in buttons.