Further improve the way we handle too large archives (#949)
This makes sure the correct message is displayed when uploading archives with
more than the allowed number of files in them. It also further improves the
handling of zip files which will explode to too large sizes (zipbombs) as these
will no never be written to disk completely. Finally it also makes sure that any
left over temporary directories are removed when extracting fails for whatever
reason.