Fix `get_from_latest_submissions` on duplicate dates (#1042)
* Fix `get_from_latest_submissions` on duplicate dates

It used to return multiple results for a user when it had multiple submissions
at exactly the same time, now it will only return a single submission, which
submission however is still undefined.

* Fix failing tests

They were failing as other functions were expecting they could do distincts and
other sorts on the returned query. Moving the selection of the newest ids to a
subquery fixes this for now, but a better fix would be nice.