Release JungleJoy.2 fixes (#1045)
* Fix z-index of local header

It was too low and this caused some UI bugs in Edge.

* Never add feedback if we are already editing feedback on that line

This fixes multiple bugs, for example you could no longer select the snippet
input field, as it would automatically focus the feedback input field again.

* Set the nice level of external commands

This makes sure our database and webserver keep priority when running long and
intensive linters/plagiarism checks. I used a niceness of 10 everywhere, but I
have no real explanation why that specific value.

* Force reload AutoTest results on load of AutoTest component

* Fix error when uploading the first submission

The `submissionsByUser[assignmentId]` might be empty, so setting
`submissionsByUser[assignmentId][userId]` fails. We now first check if it is
empty, and in that case we set an object on `submissionsByUser[assignmentId]`,
instead of setting the `userId` key.

* Fix AutoTest popover in rubric viewer

* Make sure we only load needed files in feedback overview

This fixes a bug that the page crashed when you added feedback on a file without
any feedback, as the needed `codeLines` for this file was not present. We also
reduce the amount of loading of code to the bare minimum.

* Fix reloading of submissions after dividing them

The wrong store was used, as we should use the 'submissions' store, not the
'courses' store.

* Fix bug where too many runners could be started

* Hide result list when showing single result

* Fix empty/editable items after importing a rubric

* Add pagination to the auto test result list

* Fix delete submission

* Limit max amount of runners per job

* Don't check if a runner is part of run before killing it at the broker

This is not the case when the job is started after the request, and it also
simply not necessary as the broker will check this for us.

* Retry requests when they fail in the auto test runner

* Retry tasks that do requests to the broker

* Add message below submissions table when there are no submissions

* Fix unwanted indentation in the plagiarism export

* Always fail set checkpoints if a previous one has failed