Make it possible to use CF after a final run (#1077)
* Make it possible to use CF after a final run
This is quite a large change, as it a previous assumption doesn't hold
anymore. We used to assume that a final run always takes precedence over a
continuous feedback run, however this is no longer the case. I fixed this by
adding a new prop `forceRun` to the auto test component, which forces which run
the component should use. However, this has effect through the store, as we made
the same assumption about precedence there too. In the submission page we check
if the final run has a result for the current submission, and if not we use the
continuous feedback run as run (if available of course).

* Make it impossible to start AutoTests when there are no submissions
This is important as we think these never need to be run, as they no unfinished
results. As this also really makes no sense, it seems smart to simply disable
this.