Improve auto test (#1089)
* Improve handling of multiple submissions

With this change Continuous Feedback works better with old submissions. There
are 3 changes in behavior:
- If a user hands in a new submission for an assignment with CF, all old runs
    for this user in this assignment which are not finished yet will be marked
    'skipped'.
- When a runner tries to update an old submission during a CF run an error will
    be returned to the runner, and the runner will stop running the current
    submission.
- Only the newest submissions will be send to the runners in the initial
    instructions. This makes sure they never start running an old submission.

* Add timeouts to all requests to the server

This should fix problems that happen when a instance is down for a short
time. This used to cause the runner to be stuck in a http request, as the
socket would be closed, but we would keep trying to send data on it (or
something like that). This also increases the backoff factor a bit while
retrying, this should improve the behavior when the server is down for a very
short time.

* Improve the way we handle command timeouts

This fixes 3 things:
- When a command signaled when running IO tests this was misinterpret as a
    timeout. This could cause quite some confusion.
- When a student has an infinite loop in the submitted code, this usually causes
    many commands to timeout. If you have multiple of these submissions they
    would hog the entire test server until all commands (which all timeout) are
    done. For other students this is very annoying as they can't see why their
    submissions are not tested. To fix this we now yield our core when we have a
    command times out. This causes us to go to the back of the line of students
    waiting for a core (and thus students waiting to be run), making it possible
    for other students to receive feedback again. This also has the added
    benefit that containers are now always shutdown after a command times out,
    which makes sure no processes are left running.
- This also improves the way the steps are executed, instead of passing many
    arguments we now pass an `ExecuteOptions` dataclass, which we can easily
    extend with extra info/extra callbacks.

* Try to fix deadlock when starting commands

We again experienced deadlocks when running commands in the part which waits on
the attach. This is fixed in two main ways:
- We now use `lxc.Container.signal_start` to print the null byte to stderr. This
    is a new function in our lxc library which prints a null byte to stderr
    without locking anything.
- A change to our lxc library which makes sure we do an exit when our attach
    callbacks raises an error. Our wait loop on the attach detects this, and
    stops waiting.

I've also created a test case which caused a deadlock with the old code that is
not present anymore with the new code. However, as this code really needs lxc
and should be run without the output buffering of pytest we don't run it on
Travis.

* Make sure a runner is actually started in the broker

The runner now sends a 'alive' message to the broker after it started its
template container. Using this information we can do 2 things:
- If a runner is a bit slow starting this might be just because of bad luck, but
    it might also simply be broken. This is why we now start an extra runner if
    we detect a slow startup.
- If a runner is really slow to start kill it, and create a new one if needed.

* Improve the broker panel even more

We now first show the active runners and jobs, and show the oldest runners/jobs
first. Furthermore, a legenda for runner states has been added to the sidebar.

* Add health route to the broker

* Change default sort for submissions link in the broker