Fix plagiarism overview when missing permissions on other courses (#1111)
In this case the `submissions` key is `null`, so calling `reverse` on it
fails. Furthermore, we weren't using the utils function to get the error
message, and as `response` was defined on the error object, getting `data` from
it resulted in another exception. Finally, the popover message for these cases
could be improved.