Fix race condition when creating unassigned runners (#1113)
When creating unassigned runners we created the amount of runners that was
needed at that moment. However, if a backend created a job a moment later we
didn't create a new runner, but simply assigned an unassigned runner. This
problem should now be fixed as we don't simply create unassigned runners for
jobs anymore, but always assign them as we can. We now also check if we still
have enough runners after stealing a runner, or taking an unassigned runner.