JungleJoy.4.2 hotfixes (#1144)
* Make sure rubric is visible even if user may not see AutoTest config

* Make it possible to see all AT results for a specific user

* Make sure AT will not become stuck when stopping fails

It can happen when we don't wait correctly for all processes that AT will become
stuck. This was caused by the fact that the container will not stop, but hang
when calling `cont.kill`. This commit introduces a better way to do
this. Instead of killing, we send a shutdown to the init system (`SIGPWR`), and
if it takes too long for the container to stop, we simply kill the entire worker
process and start a new one.

* Fix a race condition that could cause the argument to `time.sleep` to be negative

* Make it possible to see AutoTest results of older submissions again

If the AutoTest result is not present after loading all latest results we load
all AT results by the user of the requested submission. This will load older
AutoTest results, so we store them in another attribute then the normal
results. This makes sure we still only show the latest results on the manage
assignment page.

* Add needed new stub method

* Fix formatting

* Collapse the AutoTest config only when the run has successfully started

* Make it possible to start a AT run when there are no submissions

This makes perfect sense again, as new submissions will be automatically added
to the run when students hand-in.

* Fix comments by Olmo Kramer

* Make it possible to impersonate users

This commit makes it possible to impersonate users, when you have the correct
permissions. This feature is really useful to check why certain things do not
work correctly for students. However, as this is a really powerful feature (you
basically become an entire other user), it is important that no normal user ever
has access to this feature. This is why some LTI role mappings were changed, we
never want LTI users to be admin, as they might be admin in the LMS, that should
not mean that they are admin within CodeGrade.

* Remove "Can" from permission short description

* Fix error message

* Fix woopy CF badge

* Add comment

* Fix margin below AutoTestSets

* Hide "CF" badge when feedback is not available immediately

* Fix bug that number input for divide submissions would only accept integers