Stop loading the rubric and graders twice on the management page (#1152)
This is done by: 1) not loading the rubric in mounted and a immediate handler in
the rubric editor component, and 2) only reloading data when the `assignmentId`
changes in the assignment management component, not when the assignment
changes. Finally this will also make sure we don't wait for the graders to be
loaded before showing the page.

This makes it possible to load an initial version of the management page without
any requests. However, as this renders quite a lot of data the UX is quite bad,
it seems like the page freezes up for a second. To counteract this, I've added a
new data element `loadingInner` that only gets set to `false` after we are sure
a loader has been rendered.