Cache code in the frontend (#1138)
* Cache code in the frontend

This makes it way faster to toggle between files, as they don't have to be
redownloaded every time. This also has the added benefit of letting all things
that access code go through a common api layer. This makes it easier to change
things there if ever needed.

* Reinsert old items back in the cache when possible

* Improve speed by using more appropriate data structures.

Instead of only using a map (which stored a mapping from id to object
containing data and a sequence number) we now use a deque and map. The deque
contains the ids of the code in the order in which they were accessed (latest
first). The map contains a mapping from id to compressed data. This removes one
object per cached file, but more importantly makes it possible to remove old
items from the cache way quicker (O(n) vs (O(n * n))). The disadvantage is that
accessing a file is a tiny bit slower (as we now have to move the accessed value
to the beginning). However, in the common case (the accessed file is somewhere
at the start of the array), this doesn't matter that much.

* Let the image and pdf viewer also use the code cache store

This doesn't work for IE11, or for old versions of Edge, but we don't really
support IE11, nor versions of Edge that don't support this (12-14, which was
released in August 2016).

* Improve support for Edge and improve error message for IE

* Make sure file viewer always first show a loader before rendering