Hotfix JungleJoy.5.1 release fixes (#1165)
* Respect fixed max points in rubric and grade viewer

This was broken as the rubric store didn't have access to the assignment, and
therefore couldn't know about this option. This commit changes the content of
the rubric store, now it only stores rubric results and the rubrics themselves
are only stored on the assignment in the courses store. This fixes #1159.

* Rewrite use of store in grade viewer

* Make the progress meter in the rubric viewer reactive

* Fix race condition deleting plagiarism runs

* Fix race condition when deleting snippets

* Fix race condition when running linters after submission

* Fix LTI grade passback when used in combination with groups

It could happen that older submissions by users, which also have submissions in a
group, done individually would be passed back instead of the submission done in
the group.