Fix getting latest submissions in combination with groups
If a user had a personal submission and a group submission, the personal
submission would still be seen as a latest submission to the assignment. This
resulted in many bugs, for example in plagiarism checking and in grade
passback (although this last bug was hotfixed). With this PR the method
`Assignment.get_from_latest_submissions` returns only the latest submissions for
a user, so if a user has a group submission this is returned. Furthermore, old
personal submissions are still displayed in the front-end, but are clearly
marked to be old.