Make it possible to hand-in submission through GitHub+GitLab (#1182)
* Make it possible to hand-in submission through GitHub+GitLab

This makes it possible for students to automatically hand-in submissions by
pushing to GitHub/GitLab. Each student gets a unique url, ssh public key, and
secret, which can be used to configure a deploy key + webhook on GitHub and
GitLab. This also works for private/self-hosted GitLab instances.

This is still a work in progress, things that still need to be fixed:
- [x] When a user configures a webhook before she/he is in a group and then
    joins a group this webhook will still hand-in as the user, not the group.
- [x] The user needs more instructions for configuring the deploy key +
    webhook.
- [ ] Some visual feedback that the hook is configured correctly might be
    nice. However, it is not really clear if this is possible.
- [ ] By default we only clone the `master` branch. Students can configure
    multiple branches through the `branch` get parameter of the
    webhook. However, it would still be useful if teachers can set the default
    branch to something other than `master`.
- [x] There are no tests yet, however we really do need to test this. Something
    that might be possible is to actually run the code with GitHub+GitLab (or
    only one) webhooks by using as service like `serveo.net`.
- [x] Check if all goes well if the repository is renamed.

* Rewrite private key handling to cryptography instead of paramiko

* Fix behavior of webhooks in combination with groups

Now personal webhooks get disabled when joining a group, this is also
communicated to the user when joining a group in the form of a
warning. Furthermore, you cannot join groups that have webhooks configured for
assignments for which there is no `sourcedid` for the user that wants to join
the group. Finally, the code in the webhook now checks if at least one member of
a group can hand-in a submission, and in this case the submission is created.
* Add warning when combining webhooks with ignore files

* Add copy to clipboard popover

* Set pointerEvents instead of the cursor

* Add docs for Git submissions (#1189)