Add AutoTest output directory (#1188)
* Add AutoTest output directory

AutoTest scripts can now write files to a directory `$AT_OUTPUT`. Files written
to this directory are synced back to the CodeGrade server. This is done once per
category, and each AutoTest category starts with a fresh, empty output folder.

The files are shown in a new category in the file tree on the submission
page. The FileTree component was rewritten to allow for this as there wasn't
enough horizontal space to add an extra tab next to the teacher revision/diff
tabs. The top-level directories are now stacked on top of each other.

* Fix teacher file tree

Because we copied the file properties to a new object the "revision"
property set on the teacher file object would not be set on the file
object in the actual teacher tree. This now uses a separate set to keep
track of the files we've already seen so we don't pollute the file
objects with a "done" property.

* Redesign file tree on the submission page

* Fix new file tree when there is no revision

* Reset the revision query parameter using the submission navbar