Fix deadlock that would occur when attaching failed (#1217)
* Fix deadlock that would occur when attaching failed

This happened because the files for `stdout` and `stderr` were not closed, and
therefore the reader thread would never join. Moving the `_wait_for_attach` to
the `try` block makes sure these files are closed, so the reader thread can be
terminated.

This PR also adds a test that actually tests that when an attach fails the test
is terminated and the result is reset to `not_started` again.