Stop registering AT jobs at the broker if there are not submissions (#1220)
* Stop registering AT jobs at the broker if there are not submissions

This fixes a bug where an AT job would request a runner, but never use it if
there were no submissions (and therefore also no results to run). Now we simply
don't register the job if there are no submissions when creating an AT run. This
doesn't really fix the core of the problem, as it will probably still occur when
you create a submission, and delete it directly after.