Add support for LTI1.3
This commit moves the LTI1.1 implementation to a new file, and extends the LTI
module. Furthermore, this commit adds stubs for the only LTI1.3 library for
Python, and implements the needed classes for Flask.

DEV-60 #time 1h Split lti1.1 and lti1.3 implementations.
DEV-60 #time 7h Add stubs for pylti and implement abstract classes.