[DEV-77] Start using timezones everywhere when dealing with datetimes (#1257)
This adds a new type, `DatetimeWithTimezone` for datetimes with timezones. This
isn't an actual runtime type, but rather something that ensures in combination
with mypy that a datetime always has a `tzinfo` set, and that it can never be
removed. It doesn't force UTC, as there is no real reason for this.

This also updates moment to fix issue with `toISOString`. See https://github.com/moment/moment/pull/4391