Don't unlink password file after reading it
This can cause bugs (like the one on 2020-02-18) as the runner crash before it
got a job. In this case the runner can simply restart its service and retry to
get jobs, however when the password file is unlinked this is no longer
possible.

We do however want to make sure that runners that have received a job are no
longer eligible to receive further jobs, nor see which instances have jobs
available. To enforce this the broker should only allow runners that are in a
state before running, i.e. they haven't started processing a job yet.