Reduce locking when getting more results
It is not necessary to lock the AutoTest run when getting results, except when
it is the last call. It is necessary when it is the last call as otherwise we
might stop the runner while we make a decision based on the amount runners (when
adding work to a run).