Fix PreferenceManager
In new versions of BSV the contents of popovers is not always mounted,
but we depended on that in the preference manager, as it loads and emits
data on mount. This reworks the component so that it now accepts an
`inPopover` prop that, when given, will render only a button and shows
the preferences in a popover when the button is clicked.

This also fixes the whitespace setting, which apparently was broken
because of a race condition.