Improve speed when deleting AT runs (#1279)
With this change we stop relying on SQLAlchemy to delete all relations when
deleting an AutoTest run, and we start using the database by using the `ON
DELETE` triggers build in for foreign keys. This is way faster, as we now don't
need to load in all the data when deleting it. This should also make it possible
to delete a run while a runner is active.