Fix student count in submission list
Because we used a Set to count the number of unique names of students in
a submission list, the count would be too low when there were students
with the same name. Simply returning the length of the list of
submissions should work now that the SubmissionList is really a list of
all different users that have uploaded work to this assignment.