Move the general feedback popover into the component
This way we can have more control over when the popover is hidden, which
we use to show a confirmation to save to the user when the feedback has
changed and the popover is about to be closed. When the user clicks "no"
all changes are discarded.

This change was made in this PR rather than a new one because the
GeneralFeedbackArea would have to be changed anyway because the new BSV
popovers do not load their content when the popover is not visible, and
I did not want to do rewrite the component twice.