Fix multiple disabled role dropdowns on course members page
We compared on the user's full name to decide if the role dropdown on
the course members page should be disabled (we disable the one for the
logged in user, so they can't accidentally remove their permissions to
change user roles). But if two users have the same name, then both their
dropdowns would be disabled. This changes the check to compare user ids.