Give a better indication when an AT step is hidden
For teachers the eye icon indicating that an AT step is hidden was not
visible most of the time, because we would render either the eye or the
rotating caret when a step is expanded. Now we always render both if
both meet their conditions.

This also adds the `text-muted` class to all hidden steps to make it
even more obvious which steps are hidden.

DEV-92 #doing