[DEV-88] Do not discard changed rubric viewer when saving general feedback (#1283)
* Do not discard changed rubric viewer when saving general feedback

Whenever a rubric result changed in the store (e.g. because its
submission changed, or the AutoTest results updated it) it was set as
the current result in the RubricViewer, discarding any changes that were
made. This changes that so we now keep the changes that were made and
apply those to the store result to get the current state.