Fix prefs popover disappearing when selecting a language
The vue-multiselect component emits blur()s its list when an element is
selected, which causes popovers with trigger "click blur" to close.

This fix disables the popover temporarily while the languages list is
open, which has as a downside that when the language is open you have to
click twice to close both the list and the popover.

It may be possible to detect a click outside of the popover while the
list is open by checking if `document.activeElement` is a child of the
popover.