Improve way we handle deletion of all runners
We now simply kill the entire job and start a new one. This is less error prone
than the old behavior which was to kill every runner individually and adjust the
amount of wanted runners. The old behavior caused errors when stopping single
runner failed, and this could cause the job to become stuck.