Knoet.3 release fixes (#1293)
* Style dates for courses in HomeGrid differently

This makes it clearer that those dates are not part of the name of the course,
but instead are dates.

* Show a commit hash instead of a version for non stable deploys

This makes it easier to see which exact version you are testing.

* Fix styling of extra course loader button

* Fix sidebar assignment list gear icon top margin

* Set class on body when using CodeGrade in Edge

* Fix icons in buttons on Edge

* Fix initial LTI launch for an assignment

The assignment was not loaded in this case, which means the front-end crashed.

* Fix LTI launch for non logged in users

* Improve edge icon alignments

* Fix long usernames in group management

* Fix prefs popover disappearing when selecting a language

The vue-multiselect component emits `blur()`s its list when an element is
selected, which causes popovers with trigger "click blur" to close.

This fix disables the popover temporarily while the languages list is
open, which has as a downside that when the language is open you have to
click twice to close both the list and the popover.

It may be possible to detect a click outside of the popover while the
list is open by checking if `document.activeElement` is a child of the
popover.

* Fix wrong missing trailing newline message for some highlight langs

* Make notification description popovers hug-text

* Fix `get_assigned_grader_ids` to never yield `None`

* Fix rubric viewer lock icon on Edge

* Fix AT fixture item height in Edge

* Improve way we handle deletion of all runners

We now simply kill the entire job and start a new one. This is less error prone
than the old behavior which was to kill every runner individually and adjust the
amount of wanted runners. The old behavior caused errors when stopping single
runner failed, and this could cause the job to become stuck.

* Let `_notify_broker_of_new_job` calculate the amount of needed runners

This makes sure this value is always up to date.

* Improve webhook instructions for gitlab

* Make sure `unzip` doesn't touch the creation time of files

* Fix indentation

* Always go to the Code tab when clicking filename in plagiarism detail

* Fix new course/assignment popover widths

* Fix auto test state icon alignment

* Fix auto test set stop points input width

* Fix auto test modal sizes

* Fix HTML viewer for AutoTest files

* Fix next and previous sub popover in submission navbar

* Fix PDF viewer in safari

* Show hidden icon on AT IO substeps

* Fix HTML viewer height on safari

* Place cursor at end of feedback textarea when opening inline feedback area

* Fix content being cut off in FloatingFeedbackButton

* Add more debug output

* Fix checkboxes in plagiarism export

* Add proxy domain to csp policy

* Make it possible to get the file-url for an AutoTestOutputFile

This is needed to make the pdf viewer for AutoTestOutput files work in edge.

* Fix submission navbar border radius for students

* Fix webhook form-control heights

Co-authored-by: Olmo Kramer <olmo@codegra.de>