Fix detection of old submission by user in group (#1297)
The logic to detect that a user was viewing a personal submission of a user that
also has a group submission on the submission page was completely broken. The
worst part of this was that the computed property `groupOfCurrentUser` was
actually returning a submission, which broke many assumptions in other
components, making it possible to completely break the page.

This also makes the subcomponents check for the correct object (`models.User`)
instead of just any object.