Move loading of code to FileViewer component (#1244)
This makes each viewer component easier, as they do not need to handle async
state. This also fixes a bug where nothing would be shown on empty markdown
files.

This also fixes some errors that were not displayed in the file viewer because
the loadCode action of the code store did not rethrow errors. Furthermore,
somehow axios passed the error JSON object received from the server as an
ArrayBuffer to the defaultTransformResponse transformer, causing an error in
JSON.parse.

Co-authored-by: Olmo Kramer <olmo@codegra.de>