Add two new metrics for the health route
We now also check that jobs do not have not started tasks older than
`SLOW_STARTING_TASK_AGE` and that jobs do not have running tasks older than
`SLOW_TASK_AGE`. This commits also introduces many changes to our sqlalchemy
stubs allowing for better typed queries involving json. This change also
migrates the `JSON` columns in the broker to `JSONB`, which is faster and
the disadvantages are not really applicable to us. Finally, it also adds an
index to the `state` columns for the `Runner` and `Job` tables in the broker, as
we filter a lot on these columns so it should make a lot of queries faster.