Improve stability and monitoring of autotest (#1302)
* Make sure we catch the right exception for http tasks

When using `requests` the bottom type for their exception type hierachy is
`RequestException` not `HTTPError`. One exception that inherits from `RequestException`
but not from `HTTPError` is `ConnectionError` that gets raised when we
cannot connect to the external server (which can happen because of network
exceptions).

* Add two new metrics for the health route

We now also check that jobs do not have not started tasks older than
`SLOW_STARTING_TASK_AGE` and that jobs do not have running tasks older than
`SLOW_TASK_AGE`. This commits also introduces many changes to our sqlalchemy
stubs allowing for better typed queries involving json. This change also
migrates the `JSON` columns in the broker to `JSONB`, which is faster and
the disadvantages are not really applicable to us. Finally, it also adds an
index to the `state` columns for the `Runner` and `Job` tables in the broker, as
we filter a lot on these columns so it should make a lot of queries faster.