Fix the default chart color in dark mode
We used @color-primary as the color for the first dataset in each plot,
but since we also use that as the background in the dark theme, the data
in the plots would not be visible. This changes the color list that is
used so that it uses @text-color-dark as the color for the first dataset
in dark mode.