Add a check to see if we can reach the broker to the health route
This is useful to make sure the instance is still able and allowed to
communicate with the broker. If the instance is no longer able to do that
AutoTest runs will silently not be started, resulting in a degraded
service. This does make the health route a bit slower, but I think it is worth
it for the extra security.