Make usernames case insensitive
This makes it easier for users to login, and makes it easier for the teacher to
see who is who. The checking is done in the database, using the `citext`
extension of postgres, the nice thing about it is that while usernames are
case insensitive (both in uniqueness and in matching), their case is preserved
when displaying.

Another small tweak in this commit is that when somebody tries to login with
something that looks like an email and it fails, we notify them that their
username is probably not their email.