Add sentry integration to CodeGrade (#1314)
* Add sentry integration to CodeGrade

This makes it possible to do error reporting to our self hosted sentry
instance (although any sentry instance should work). It makes it possible to log
to sentry by passing `report_to_sentry` to a log call, it also automatically
logs every 500 error, and any call to `logger.error` in AutoTest.

This also adds sentry to our front-end, where it logs any unhanded
exception. Users can opt-out by blocking the sentry bundle, which we currently
handle well.

* Fix stuck linters

Now that `TYPE_CHECKING` is no longer set while building the docs (that was
stupid idea of mine for just a tiny bit better docs), as this is no longer
possible because `sentry` also uses it to conditionally do stuff that crashes
art runtime, we shouldn't setup the periodic tasks when we have no celery config
set, as this causes the building of the docs to hang forever.