Fix AutoTest error caused by a race condition
On the assignment management page there are two polls for AT results,
both updating the AutoTestRun model in different ways, sometimes causing
an error when viewing a single result (A) when the user of that result
creates a new submission (B).

When this happens, the last polling check may still find result A in the
AutoTestRun.results array (which is the list of latest results for the
assignment). Because this result is final (it's state is "skipped"
because there is a new submission), polling for the latest result stops.

Then, after a few seconds, when the entire run is polled again by the
background AutoTest component, the server returns the new result for the
student, clearing the previous result from the list of latest results.
But because polling has stopped, the old result is never retrieved
again. And because it was still in the "latest results" list on the last
poll, it has never been loaded into the AutoTestRun.resultsByUser
mapping. So now the result will always be undefined.

This commit makes a few changes:

- Prevent an error while the result is undefined.
- Reload the result if we detect that we should be showing a result, but
  none seems to be loaded.